testando  do while pra mostrar o menu

+++ src/Models/ComissarioBordo.java
@@ -9,7 +9,7 @@ public class ComissarioBordo extends Pessoas {
 
     }
 
-    public ComissarioBordo(String nome, int cpf, String telefone, int idCarteiraComissario, String funcaoVoo) {
+    public ComissarioBordo(String nome, long cpf, String telefone, int idCarteiraComissario, String funcaoVoo) {
         super(nome, cpf, telefone);
         this.idCarteiraComissario = idCarteiraComissario;
         this.funcaoVoo = funcaoVoo;
